refactor: move the data directory layout and version gate into raft-config
# Summary

Creating a node's data directories, reading and writing `df_meta/VERSION`,
and refusing to start on data this build is too new for are permanent
concerns. They lived on `OnDisk` in the compatibility crate, so nothing
could ask "what version is this data directory" without depending on the
V003 upgrader.

# Details

`raft-config` gains a `data_dir` module holding `ensure_dirs`,
`header_path`, `load_header`, `write_header` and `assert_compatible`. It
already owned `RaftConfig`, `DataVersion` and the snapshot path
computation, so this is the same concern rejoining itself rather than a
new responsibility.

`OnDisk` keeps its public shape and delegates, so the downstream entry
point `OnDisk::open(config).upgrade()` is unaffected. What stays behind
is exactly the V003 behavior that had accumulated inside the live path:
migrating a header out of the legacy sled tree, defaulting a
header-less directory to V003, and the upgrade state machine.

The two tests covering the moved code — a malformed version file, and
the too-old-data panic — move with it.
